requirements.txt
fastapi==0.110.1
flake8==7.1.1
uvicorn==0.18.3
SQLAlchemy==2.0.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
passlib[bcrypt]==1.7.4
pyjwt==2.4.0
pytest==7.1.2
pytest-asyncio==0.18.3
httpx==0.23.0
passlib==1.7.4
python-multipart==0.0.5
black==24.8.0
isort==5.13.2
